attempt to take audio functions out of main cpp code, issue currently with pipefds
