+++ Urho3D/splash/audioIn.h
@@ -1,0 +1,1 @@
+int audioIn();
