--- Urho3D/splash/lol.h
@@ -1,1 +1,0 @@
-int lol();
