+++ Urho3D/splash_build/CMakeFiles/Splash.dir/CXX.includecache
@@ -1276,103 +1276,13 @@ Urho3D/Resource/XMLFile.h
 Urho3D/IO/Log.h
 -
 
-/home/pi/sound-pirates/Urho3D/splash/fft.cpp
-RtAudio.h
-/home/pi/sound-pirates/Urho3D/splash/RtAudio.h
-fftw3.h
--
-iostream
--
-cstdlib
--
-math.h
--
-SFML/Graphics.hpp
--
-cstring
--
-stdio.h
--
-SFML/Audio.hpp
--
-vector
--
-complex
--
-string
--
-lol.h
-/home/pi/sound-pirates/Urho3D/splash/lol.h
+/home/pi/sound-pirates/Urho3D/splash/audioIn.h
 
 /home/pi/sound-pirates/Urho3D/splash/fft.h
 
 /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
 
-/home/pi/sound-pirates/Urho3D/splash/lol.h
-
 /home/pi/sound-pirates/Urho3D/splash/playNote.h
-
-/home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp
-RtAudio.h
-/home/pi/sound-pirates/Urho3D/splash/RtAudio.h
-fftw3.h
--
-iostream
--
-cstdlib
--
-math.h
--
-SFML/Graphics.hpp
--
-cstring
--
-stdio.h
--
-SFML/Audio.hpp
--
-vector
--
-complex
--
-string
--
-lol.h
-/home/pi/sound-pirates/Urho3D/splash/lol.h
-fft.h
-/home/pi/sound-pirates/Urho3D/splash/fft.h
-playNote.h
-/home/pi/sound-pirates/Urho3D/splash/playNote.h
-instructionsStatements.h
-/home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
-sys/types.h
--
-stdlib.h
--
-fcntl.h
--
-string
--
-memory
--
-fstream
--
-sstream
--
-iostream
--
-stdio.h
--
-unistd.h
--
-poll.h
--
-thread
--
-chrono
--
-signal.h
--
 
 /home/pi/sound-pirates/Urho3D/splash/processBuffer.h
 
@@ -1401,8 +1311,8 @@ complex
 -
 string
 -
-lol.h
-/home/pi/sound-pirates/Urho3D/splash/lol.h
+audioIn.h
+/home/pi/sound-pirates/Urho3D/splash/audioIn.h
 fft.h
 /home/pi/sound-pirates/Urho3D/splash/fft.h
 playNote.h
