+++ Urho3D/splash_build/CMakeFiles/Splash.dir/depend.internal
@@ -6,30 +6,30 @@ CMakeFiles/Splash.dir/RtAudio.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
 CMakeFiles/Splash.dir/fft.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+ /home/pi/sound-pirates/Urho3D/splash/audioIn.h
  /home/pi/sound-pirates/Urho3D/splash/fft.cpp
- /home/pi/sound-pirates/Urho3D/splash/lol.h
 CMakeFiles/Splash.dir/instructionsStatements.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+ /home/pi/sound-pirates/Urho3D/splash/audioIn.h
  /home/pi/sound-pirates/Urho3D/splash/fft.h
  /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.cpp
- /home/pi/sound-pirates/Urho3D/splash/lol.h
  /home/pi/sound-pirates/Urho3D/splash/playNote.h
 CMakeFiles/Splash.dir/playNote.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+ /home/pi/sound-pirates/Urho3D/splash/audioIn.h
  /home/pi/sound-pirates/Urho3D/splash/fft.h
- /home/pi/sound-pirates/Urho3D/splash/lol.h
  /home/pi/sound-pirates/Urho3D/splash/playNote.cpp
  /home/pi/sound-pirates/Urho3D/splash/sinewave.h
 CMakeFiles/Splash.dir/processBuffer.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
+ /home/pi/sound-pirates/Urho3D/splash/audioIn.h
  /home/pi/sound-pirates/Urho3D/splash/fft.h
  /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
- /home/pi/sound-pirates/Urho3D/splash/lol.h
  /home/pi/sound-pirates/Urho3D/splash/playNote.h
  /home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp
 CMakeFiles/Splash.dir/realtime-beat.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
- /home/pi/sound-pirates/Urho3D/splash/lol.h
+ /home/pi/sound-pirates/Urho3D/splash/audioIn.h
  /home/pi/sound-pirates/Urho3D/splash/realtime-beat.cpp
 CMakeFiles/Splash.dir/splash.cpp.o
  /home/pi/Urho3D_build/include/Urho3D/Audio/Audio.h
@@ -182,9 +182,9 @@ CMakeFiles/Splash.dir/splash.cpp.o
  /home/pi/sound-pirates/Urho3D/splash/RtAudio.h
  /home/pi/sound-pirates/Urho3D/splash/Sample.h
  /home/pi/sound-pirates/Urho3D/splash/Sample.inl
+ /home/pi/sound-pirates/Urho3D/splash/audioIn.h
  /home/pi/sound-pirates/Urho3D/splash/fft.h
  /home/pi/sound-pirates/Urho3D/splash/instructionsStatements.h
- /home/pi/sound-pirates/Urho3D/splash/lol.h
  /home/pi/sound-pirates/Urho3D/splash/playNote.h
  /home/pi/sound-pirates/Urho3D/splash/processBuffer.h
  /home/pi/sound-pirates/Urho3D/splash/splash.cpp
